Started output device review

+++ src/MF_Segment/LedSegment.h
@@ -1,8 +1,10 @@
 #pragma once
+#include <stdint.h>
+#include "MFSegments.h"
 
 namespace LedSegment
 {
-void Add(int dataPin, int csPin, int clkPin, int numDevices, int brightness);
+void Add(uint8_t dataPin, uint8_t csPin, uint8_t clkPin, uint8_t numDevices, uint8_t brightness);
 void Clear();
 void PowerSave(bool state);
 void OnInitModule();
