+++ src/MF_Stepper/Stepper.h
@@ -1,8 +1,10 @@
 #pragma once
+#include <stdint.h>
+#include "MFStepper.h"
 
 namespace Stepper
 {
-void Add(int pin1, int pin2, int pin3, int pin4, int btnPin1);
+void Add(uint8_t pin1, uint8_t pin2, uint8_t pin3, uint8_t pin4, uint8_t btnPin1);
 void Clear();
 void OnSet();
 void OnReset();
